Added a test demonstrating send value duplication in the current model.
The error happens when multiple receivers obtain the same value from a channel on account of  a second sender overwriting the first sender's stored value before it is pulled out.

We can resolve the error this example, overload, demonstrates by applying the buffered channel change we've discussed so each sender has a reserved place to hold what they want to store.