Adapted for ableC change from [Qualifier] -> Qualifiers.
Somewhere in this change, two things happened: a grammar was added to the include paths that had 'artifact' as its name (for the second time) so we've changed the go concurrency artifact to no longer be 'artifact', despite the standard, as the particular bug of 'the wrong grammr is being built' is hard to track down!

The second change is that ableC will now cast chars to doubles, so the error case we had that broke there has been mo,ved.